requirements.txt
# === Core ===
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.1.0
openai>=1.40.0
python-dotenv>=1.0.1

# === Linting & Static Analysis ===
pylint>=3.2.6
flake8>=7.1.1
bandit>=1.7.9

cpplint>=1.6.1

# === Testing & Dynamic Analysis ===
pytest>=8.3.3
pytest-cov>=5.0.0

# === Misc ===
requests>=2.32.3
tqdm>=4.66.5
rich>=13.9.2
